Rename rest of the files inside resutruct_poc
