+++ src/restruct_poc/solver/state.hpp
